Master create xml docs on current code (#4)
* 💡 Add docs on the services currently implemented

* 💡 More documentation stuff and typo fixes

* ♻️ Help definition attribute

This refactored code is cleaner and more readable. It uses the Path class for file path manipulations, which is safer and more reliable than manually splitting and combining paths. It also uses string.IsNullOrEmpty for string emptiness checks, which is more idiomatic in C#. The use of var has been reduced to make the types of variables clearer. The _pageHelpFileName field has been converted to a property, which is more idiomatic in C#. The nameof operator is used instead of hardcoding the string names of properties. The is pattern matching is used for type checks, which is more idiomatic in C#.